refactor: перенести приложение calc наружу пакета проекта
django-admin startapp кладёт приложение рядом с пакетом проекта, а не внутрь
него. Приложение переехало из hello_django/calc в calc, шаблоны из
hello_django/templates/calc в calc/templates/calc.

Метка приложения не изменилась: django берёт её из последнего компонента name,
то есть и hello_django.calc, и calc дают метку calc. Поэтому миграции и
закоммиченная db.sqlite3 остаются валидными, makemigrations --check это
подтверждает.
