+++ calc/apps.py
@@ -3,4 +3,4 @@ from django.apps import AppConfig
 
 class CalcConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'hello_django.calc'
+    name = 'calc'
