+++ calc/views.py
