+++ hello_django/urls.py
@@ -3,5 +3,5 @@ from hello_django import views
 
 urlpatterns = [
     path('', views.IndexView.as_view(), name='index'),
-    path('calc/', include('hello_django.calc.urls')),
+    path('calc/', include('calc.urls')),
 ]
