+++ pyproject.toml
@@ -27,8 +27,8 @@ extend-exclude = ["migrations"]
 
 [tool.ruff.lint.per-file-ignores]
 # Django boilerplate stubs
-"hello_django/calc/admin.py" = ["F401"]
-"hello_django/calc/tests.py" = ["F401"]
+"calc/admin.py" = ["F401"]
+"calc/tests.py" = ["F401"]
 
 [tool.pytest.ini_options]
 DJANGO_SETTINGS_MODULE = "hello_django.settings"
